Corrected a typo

+++ src/main.py
@@ -53,7 +53,7 @@ from datasets.main import load_dataset
               help='Number of workers for data loading. 0 means that the data will be loaded in the main process.')
 @click.option('--normal_class', type=int, default=0,
               help='Specify the normal class of the dataset (all other classes are considered anomalous).')
-@click.option('--beta', type=int, default=0.05,
+@click.option('--beta', type=float, default=0.05,
               help='beta parameter, multiplier of the ae_loss')
 def main(dataset_name, net_name, xp_path, data_path, load_config, load_model, objective, nu, device, seed,
          optimizer_name, lr, n_epochs, lr_milestone, batch_size, weight_decay, pretrain, ae_optimizer_name, ae_lr,
